requirements-docs.txt
app_version>=0.1.1
django>=1.3
django-appconf
django-override-settings
PyYAML
sphinx
south
sphinxcontrib-napoleon
